Switched timing to glfw to improve performance.

+++ src/game_engine/game_object/components/mod.rs
@@ -4,7 +4,7 @@ pub use test_component::TestComponent;
 use super::GameObject;
 
 pub trait Component {
-    fn update(&mut self, _delta_time: f32, _owner: &GameObject) {}
-    fn fixed_update(&mut self, _delta_time: f32, _owner: &GameObject) {}
-    fn render(&mut self, _delta_time: f32, _owner: &GameObject) {}
+    fn update(&mut self, _delta_time: f64, _owner: &GameObject) {}
+    fn fixed_update(&mut self, _delta_time: f64, _owner: &GameObject) {}
+    fn render(&mut self, _delta_time: f64, _owner: &GameObject) {}
 }
